Aggressively relocating classes to streamline the main package.
