+++ jave-core-test-java11/src/test/java/ws/schild/jave/MultimediaObjectTest.java
@@ -23,6 +23,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import org.junit.jupiter.api.Test;
 
+import ws.schild.jave.info.MultimediaInfo;
+
 /**
  *
  * @author a.schild
