+++ jave-core-test/src/test/java/ws/schild/jave/EncoderTest.java
@@ -23,6 +23,9 @@ import java.util.concurrent.TimeUnit;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 
+import ws.schild.jave.encode.AudioAttributes;
+import ws.schild.jave.encode.EncodingAttributes;
+
 /**
  *
  * @author a.schild
