+++ jave-core-test/src/test/java/ws/schild/jave/EncodingAttributesTest.java
@@ -22,6 +22,11 @@ import java.io.File;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 
+import ws.schild.jave.encode.AudioAttributes;
+import ws.schild.jave.encode.EncodingAttributes;
+import ws.schild.jave.encode.VideoAttributes;
+import ws.schild.jave.info.VideoSize;
+
 /**
  *
  * @author a.schild
