+++ jave-core-test/src/test/java/ws/schild/jave/MultimediaObjectTest.java
@@ -24,6 +24,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import org.junit.jupiter.api.Test;
 
+import ws.schild.jave.info.MultimediaInfo;
+
 /**
  *
  * @author a.schild
