+++ jave-core-test/src/test/java/ws/schild/jave/PListener.java
@@ -8,6 +8,8 @@ package ws.schild.jave;
 import java.util.LinkedList;
 import java.util.List;
 
+import ws.schild.jave.info.MultimediaInfo;
+
 /**
  *
  * @author a.schild
