+++ jave-core/src/main/java/ws/schild/jave/encode/VideoAttributes.java
@@ -16,10 +16,15 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ws.schild.jave;
+package ws.schild.jave.encode;
 
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.Optional;
+
+import ws.schild.jave.Encoder;
+import ws.schild.jave.filters.VideoFilter;
+import ws.schild.jave.info.VideoSize;
 
 /**
  * Attributes controlling the video encoding process.
