+++ jave-core/src/main/java/ws/schild/jave/filters/Color.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package ws.schild.jave;
+package ws.schild.jave.filters;
 
 /**
  *
