+++ jave-core/src/main/java/ws/schild/jave/info/VideoInfo.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ws.schild.jave;
+package ws.schild.jave.info;
 
 /**
  * Instances of this class report informations about a video stream that can be
