+++ jave-core/src/main/java/ws/schild/jave/info/VideoSize.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ws.schild.jave;
+package ws.schild.jave.info;
 
 import java.io.Serializable;
