+++ jave-core/src/main/java/ws/schild/process/ffmpeg/DefaultFFMPEGLocator.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ws.schild.jave;
+package ws.schild.process.ffmpeg;
 
 import java.io.File;
 import java.io.IOException;
