+++ jave-core/src/test/java/ws/schild/jave/UtilsTest.java
@@ -6,6 +6,9 @@
 package ws.schild.jave;
 
 import org.junit.jupiter.api.Test;
+
+import ws.schild.utils.Utils;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
